simple fs-based schema registry
This expands upon @gwicke's prototype to remove the hard-coded test schema
in favor of one that is stored on the file-system.  Additional schemas can
be added as well provided that the filename corresponds to the configured
topic (i.e. for a topic of "edits", the schema will be loaded from
schemas/edits.js).

This is obviously Not Good Enough, but can be easily expanded upon as
needed; We are after all, in MVP mode.

Bug: https://phabricator.wikimedia.org/T114443